test refactor - pulling midpoint function out
